refactor: [#18] 3차 PR 반영
