+++ src/SubCard.js
@@ -6,10 +6,19 @@ import CategoryTitle from './components/CategoryTitle';
 const SubCard = ({ dishData, title }) => {
   const mainDatas = dishData.slice(0, 9);
   const cardDataLen = dishData.length;
-  const cardNum = 2 < cardDataLen && cardDataLen < 4 ? cardDataLen : 4;
+  const originalShowingCardNum = 4;
+  const specialShowingCardNum = 3;
+  const showingSlideCardNum =
+    cardDataLen === specialShowingCardNum
+      ? cardDataLen
+      : originalShowingCardNum;
 
   return (
-    <CardContainer cardInfos={mainDatas} hasButton={true} cardNum={cardNum}>
+    <CardContainer
+      cardInfos={mainDatas}
+      hasButton={true}
+      showingSlideCardNum={showingSlideCardNum}
+    >
       <CategoryTitle title={title} />
     </CardContainer>
   );
